Simplify route node-disjointness in model
We do not need arriving/leaving zone logic with an extra bridging variable
that is at most 1. This logic is based on an 1-1 flow construction from
the graph setting. For our linear programming formulation, it is entirely
equivalent to leave out the extra variable and say that the sum of all
incoming values is at most 1.