Add the first kind of symmetry-breaking constraints
We say that a tour must not be used if the tour by the
previous vehicle is not used.

This is not new, already mentioned in

Claudia Archetti, Natashia Boland, M. Grazia Speranza (2017)
A Matheuristic for the Multivehicle Inventory Routing Problem.
INFORMS Journal on Computing 29(3):377-387.
http://dx.doi.org/10.1287/ijoc.2016.0737

More discussed in

Claudia Archetti, Nicola Bianchessi, Stefan Irnich and M. Grazia Speranza
Formulations for an inventory routing problem
Intl. Trans. in Op. Res. 21 (2014) 353–374
DOI: 10.1111/itor.12076

And well-researched in

Adulyasak, Y., Cordeau, J.-F., Jans, R., 2013b.
Formulations and branch-and-cut algorithms for multivehicle production
and inventory routing problems.
INFORMS Journal on Computing. doi:10.1287/ijoc.2013.0550.